Move IsNotMountPoint to windows to keep it as it.
Move IsNotMountPoint to windows. Support only Linux. Throw error
on all other platforms.
